# openapi.yaml
openapi: 3.1.0
info:
  title: pdf_oxide_api
  version: 0.1.0
  description: >
    Stateless single-shot REST service wrapping the pdf_oxide library.
    No PDF data is persisted server-side. PDF in -> result out.


    Every data endpoint accepts the PDF in one of three encodings:
    multipart/form-data (a `file` part), application/json (a base64 PDF in
    `pdf_base64`, or `pdfs_base64` for merge), or a raw PDF body
    (`application/pdf`). Endpoint parameters (`pages`, `detect_headings`,
    `flatten`) are sent as top-level multipart text parts, top-level JSON keys,
    or URL query parameters — they are NOT nested under an `options` object.


    This hand-authored document is the source-of-truth contract; the running
    service serves an identical copy at /openapi.json and /openapi.yaml.
  license:
    name: MIT OR Apache-2.0
servers:
  - url: /
...
paths:
...
  /v1/docs/metadata:
    post:
      tags: [documents]
      summary: Read document metadata
      operationId: metadata
      requestBody: { $ref: '#/components/requestBodies/PdfOnly' }
      responses:
        '200':
          description: Metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetadataResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
  /v1/pipeline:
    post:
      tags: [pipeline]
      summary: Chain operations over one in-memory PDF
      operationId: pipeline
      description: >
        Runs an ordered list of operations against a single PDF. `fill` is a
        transform (PDF -> PDF). The data ops `extract_text`, `extract_markdown`,
        `extract_html`, `metadata`, and `page_info` are terminal and must be the
        last step. A pipeline ending in `fill` returns the resulting PDF; a
        pipeline ending in a data op returns that op's JSON.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, operations]
              properties:
                file: { type: string, format: binary }
                operations: { $ref: '#/components/schemas/Operations' }
          application/json:
            schema:
              type: object
              required: [pdf_base64, operations]
              properties:
                pdf_base64: { type: string, contentEncoding: base64 }
                operations: { $ref: '#/components/schemas/Operations' }
      responses:
        '200':
          description: Final PDF (pipeline ended in a transform) or final JSON (ended in a data op)
          content:
            application/pdf: { schema: { type: string, format: binary } }
            application/json: { schema: {} }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
components:
  requestBodies:
    PdfOnly:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            required: [file]
            properties:
              file: { type: string, format: binary }
        application/json:
          schema:
            type: object
            required: [pdf_base64]
            properties:
              pdf_base64: { type: string, contentEncoding: base64 }
        application/pdf:
          schema: { type: string, format: binary }
...
  responses:
    Error:
      description: Error envelope (RFC 9457 compatible)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    PagesParam:
      type: string
      description: >
        1-based page selector. Absent means all pages. A single number ("2")
        selects that page; a string like "1-3,5" selects those pages/ranges.
      examples: ["1-3,5"]
    FieldMap:
      type: object
      additionalProperties:
        oneOf:
          - { type: string }
          - { type: number }
          - { type: boolean }
          - { type: array, items: { type: string } }
          - { type: 'null' }
      description: >
        AcroForm field name -> value. Strings/numbers/booleans set a value, an
        array of strings sets a multi-choice selection, null clears the field.
        UTF-8 / CJK values are stored verbatim. In multipart this part is a
        JSON-object string.
    Operations:
      type: array
      minItems: 1
      maxItems: 16
      description: >
        Ordered pipeline steps. In multipart this part is a JSON-array string.
      items:
        type: object
        required: [op]
        properties:
          op:
            type: string
            enum: [fill, extract_text, extract_markdown, extract_html, metadata, page_info]
          fields: { $ref: '#/components/schemas/FieldMap' }
          flatten: { type: boolean, default: false }
          pages: { $ref: '#/components/schemas/PagesParam' }
          detect_headings: { type: boolean, default: true }
        additionalProperties: true
...
    MetadataResponse:
      type: object
      properties:
        title: { type: [string, "null"] }
        author: { type: [string, "null"] }
        subject: { type: [string, "null"] }
        keywords: { type: [string, "null"] }
        producer: { type: [string, "null"] }
        creation_date: { type: [string, "null"] }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - PAYLOAD_TOO_LARGE
                - PDF_ENCRYPTED
                - MALFORMED_PDF
                - UNSUPPORTED_PDF_FEATURE
                - PDF_PROCESSING_FAILED
                - LIMIT_EXCEEDED
                - NOT_READY
                - INTERNAL_ERROR
            message: { type: string }
            details: { type: [string, "null"] }
            request_id: { type: [string, "null"] }